backend/app/config.py
"""
Configuration settings for the chat application.
Contains database URLs, secret keys, and other settings.
"""

import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Database configuration
# Using SQLite for simplicity, can be changed to MySQL
DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./chat.db")

# JWT Configuration
SECRET_KEY = os.getenv("SECRET_KEY", "your-secret-key-here-change-in-production")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

# CORS Configuration
ALLOWED_ORIGINS = [
    "http://localhost:8000",
    "http://127.0.0.1:8000",
    "http://localhost:5500",  # Live server
    "http://127.0.0.1:5500",
]

# Admin credentials (in production, use database)
ADMIN_USERNAME = os.getenv("ADMIN_USERNAME", "admin")
ADMIN_PASSWORD = os.getenv("ADMIN_PASSWORD", "admin123")